+++ frontend/src/App.tsx
@@ -5,9 +5,9 @@ import LoginPage from "./pages/Login";
 import RegisterPage from "./pages/Register";
 import HomePage from "./pages/Home";
 import { tokenStore } from "./auth/token";
-import UploadPage from "./pages/Upload";
-import UploadSetupPage from "./pages/UploadSetup";
-import ConsentPage from "./pages/Consent";
+import UploadPage from "./pages/upload/upload/UploadPage";
+import UploadSetupPage from "./pages/upload/setup/SetupPage";
+import ConsentPage from "./pages/upload/consent/ConsentPage";
 import ProjectsPage from "./pages/Projects";
 import InsightsPage from "./pages/InsightsPage";
 import OutputsPage from "./pages/Outputs";
