+++ frontend/src/pages/upload/UploadShared.css
@@ -1,0 +1,20 @@
+.wizardPlaceholderCard {
+  max-width: 1000px;
+  background: var(--upload-bg);
+  border: 0;
+  border-radius: 10px;
+  padding: 22px;
+}
+
+.wizardPlaceholderTitle {
+  margin: 0 0 10px;
+  font-size: 28px;
+  line-height: 1.2;
+}
+
+.wizardPlaceholderText {
+  margin: 0;
+  font-size: 15px;
+  line-height: 1.35;
+  color: #222;
+}
