+++ frontend/src/pages/upload/upload/UploadPage.css
@@ -1,24 +1,3 @@
-.wizardPlaceholderCard {
-  max-width: 1000px;
-  background: var(--upload-bg);
-  border: 0;
-  border-radius: 10px;
-  padding: 22px;
-}
-
-.wizardPlaceholderTitle {
-  margin: 0 0 10px;
-  font-size: 28px;
-  line-height: 1.2;
-}
-
-.wizardPlaceholderText {
-  margin: 0;
-  font-size: 15px;
-  line-height: 1.35;
-  color: #222;
-}
-
 .uploadStageProgress {
   margin-bottom: 18px;
 }
